requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
drf-nested-routers==0.91
es-common==0.7.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.8
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.0.1
zipp==3.1.0
